Enhance task and UI functionality
- Added a check to ensure the stop signal file is only removed if it exists, preventing potential errors.
- Removed obsolete button classes for refresh and data selection to streamline the UI.
- Introduced a new constant for reward size in data headers.
- Updated UI to display session time, mean wait time, and total rewards, improving user feedback during experiments.
- Adjusted data plotting logic to accommodate the new reward size data, enhancing the analysis capabilities.

+++ krave/experiment/task.py
@@ -289,7 +289,8 @@ class Task:
                         print("Received STOP signal from UI. Shutting down.")
                         self.running = False
                         self.ending_code = "manual"
-                        os.remove(stop_signal_path)
+                        if os.path.exists(stop_signal_path):
+                            os.remove(stop_signal_path)
 
             if self.state == states.IN_BACKGROUND:
                 self.handle_background_events()
